# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "adminDLV API",
        "version": "1.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Danh sách user",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Danh sách user"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Tạo mới user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Chi tiết user",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User detail"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Cập nhật user",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "responses": {
                    "200": {
                        "description": "User updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Xóa user",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "responses": {
                    "200": {
                        "description": "User deleted"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}